Ejecuta una funcion dada como parametro en la cual recorremos una matriz.

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -7,7 +7,7 @@
 */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	unsigned int iterador = 0;
+	unsigned long iterador = 0;
 
 	if (!array || !size)
 	{
